+++ docker/rootfs/app/schedules/daily.sh
@@ -4,5 +4,4 @@ python3 /app/manga-dlp.py \
     --path /app/downloads \
     --read /app/mangas.txt \
     --chapters all \
-    --wait 2 \
-    --warn
+    --wait 2